+++ src/lib/CallControl/index.js
@@ -90,7 +90,7 @@ export default function CallControlDefault() {
     <div className='row'>
 
       <div className="docs-example docs-example--spacing">
-        <h3> 
+        <h3>
           <p><code className="small">type=(microphone-muted)</code></p>
           <p><code className="small">size=(20)</code></p>
           <p><code className="small">iconSize=(10)</code></p>
@@ -104,7 +104,7 @@ export default function CallControlDefault() {
       </div>
 
       <div className="docs-example docs-example--spacing">
-        <h3> 
+        <h3>
           <p><code className="small">type=(microphone-muted)</code></p>
           <p><code className="small">size=(40)</code></p>
           <p><code className="small">iconSize=(16)</code></p>
@@ -118,7 +118,7 @@ export default function CallControlDefault() {
       </div>
 
       <div className="docs-example docs-example--spacing">
-        <h3> 
+        <h3>
           <p><code className="small">type=(microphone-muted)</code></p>
         </h3>
         <CallControl
@@ -126,7 +126,7 @@ export default function CallControlDefault() {
           ariaLabel='For the Win'
         />
       </div>
-      
+
     </div>
   );
 }
@@ -148,7 +148,19 @@ export default function CallControlActive() {
     <div className='row'>
 
       <div className="docs-example docs-example--spacing">
-        <h3> 
+        <h3>
+          <p><code className="small">active=(true)</code></p>
+        </h3>
+        <CallControl
+          type='microphone-muted'
+          active
+          onClick={() => {}}
+          ariaLabel='For the Win'
+        />
+      </div>
+      <div className="docs-example docs-example--spacing cui--contrast">
+        (With Contrast)
+        <h3>
           <p><code className="small">active=(true)</code></p>
         </h3>
         <CallControl
@@ -180,7 +192,19 @@ export default function CallControlDisabled() {
      <div className='row'>
 
       <div className="docs-example docs-example--spacing">
-        <h3> 
+        <h3>
+          <p><code className="small">disabled=(true)</code></p>
+        </h3>
+        <CallControl
+          type='microphone-muted'
+          disabled
+          onClick={() => {}}
+          ariaLabel='For the Win'
+        />
+      </div>
+      <div className="docs-example docs-example--spacing cui--contrast">
+        (With Contrast)
+        <h3>
           <p><code className="small">disabled=(true)</code></p>
         </h3>
         <CallControl
@@ -212,7 +236,19 @@ export default function CallControlCancel() {
       <div className='row'>
 
       <div className="docs-example docs-example--spacing">
-        <h3> 
+        <h3>
+          <p><code className="small">type=(cancel)</code></p>
+        </h3>
+        <CallControl
+          type='cancel'
+          active
+          onClick={() => {}}
+          ariaLabel='For the Win'
+        />
+      </div>
+      (With Contrast)
+      <div className="docs-example docs-example--spacing cui--contrast">
+        <h3>
           <p><code className="small">type=(cancel)</code></p>
         </h3>
         <CallControl