+++ src/lib/SideNav/index.js
@@ -6,10 +6,9 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Icon } from '@collab-ui/react';
+import { Icon, ListItemSection } from '@collab-ui/react';
 
 class SideNav extends React.Component {
-
   state = {
     expanded: this.props.expanded
   };
@@ -21,24 +20,55 @@ class SideNav extends React.Component {
   }
 
   render() {
-    const { children, navSectionTitle, topMenu, expandable, className } = this.props;
+    const { children, navSectionNode, navSectionTitle, topMenu, expandable, className } = this.props;
+    const { expanded } = this.state;
     const topMenuTitle = topMenu ? 'top' : 'sub';
-    const expandedStatus = (!expandable || this.state.expanded) ? 'expanded' : 'collapse';
-    const navSectionTitleText= navSectionTitle &&
-      <h3
-        className={`cui-side-nav__title--${topMenuTitle}`}
-      >
-        <button
-          className='cui-side-nav__button'
-          onClick={() => {expandable ? this.handleNavToggle() : false;}}>
-          {navSectionTitle}
-          {expandable && expandedStatus === 'collapse' && <Icon name='arrow-down_16' />}
-          {expandable && expandedStatus === 'expanded' && <Icon name='arrow-up_16' />}
-        </button>
-      </h3>;
+
+    const navSectionTitleText= navSectionTitle 
+      ? (
+        <h3
+          className={`cui-side-nav__title--${topMenuTitle}`}
+        >
+          <button
+            className='cui-side-nav__button'
+            onClick={() => {expandable ? this.handleNavToggle() : false;}}>
+            {navSectionTitle}
+            {
+              expandable && 
+                (
+                  expanded
+                  ? <Icon name='arrow-up_16' />
+                  : <Icon name='arrow-down_16' />
+                )
+            }
+          </button>
+        </h3>
+      ) 
+      : navSectionNode && React.cloneElement(
+        navSectionNode,
+        {
+          ...expandable && { onClick: () => this.handleNavToggle() } 
+        },
+        [
+          ...navSectionNode.props.children,
+          ...expandable && 
+            <ListItemSection position='right' key='side-nav--expand-section'>
+              {
+                expanded
+                ? <Icon name='arrow-down_12' />
+                : <Icon name='arrow-up_12' />
+              }
+            </ListItemSection>
+          ]
+      );
+      
 
     return (
-      <div className={`cui-side-nav ${className} cui-side-nav--${expandedStatus}`}>
+      <div className={
+        'cui-side-nav' +
+        ` cui-side-nav--${(!expandable || expanded) ? 'expanded' : 'collapse'}` +
+        `${className && ` ${className}` || ''}`
+      }>
         {navSectionTitleText}
         {children}
       </div>
@@ -55,6 +85,8 @@ SideNav.propTypes = {
   expandable: PropTypes.bool,
   /** @prop Set navigation expanded or collapsed | false */
   expanded: PropTypes.bool,
+  /** @prop Navigation Section Node | null */
+  navSectionNode: PropTypes.node,
   /** @prop Title for the side navigation | '' */
   navSectionTitle: PropTypes.string,
   /** @prop Sets side navigation as the top level | false */
@@ -66,6 +98,7 @@ SideNav.defaultProps = {
   className: '',
   expandable: false,
   expanded: false,
+  navSectionNode: null,
   navSectionTitle: '',
   topMenu: false,
 };